misc and misc data processing refactor
